+++ src/components/Common/Toast/ToastContainer/style.ts
@@ -1,5 +1,7 @@
 import styled from '@emotion/styled';
 
+import { getPositionStyle } from '@/utils/getPositionStyle';
+
 interface ToastWrapperProps {
   position?: string;
 }
@@ -10,33 +12,5 @@ export const ToastWrapper = styled.div<ToastWrapperProps>`
   display: flex;
   flex-direction: column;
   gap: 1rem;
-  ${({ position }) => {
-    switch (position) {
-      case 'top-left':
-        return `
-          top: 3rem;
-          left: 0;
-        `;
-      case 'top-right':
-        return `
-          top: 3rem;
-          right: 0;
-        `;
-      case 'bottom-left':
-        return `
-          bottom: 5rem;
-          left: 0;
-        `;
-      case 'bottom-right':
-        return `
-          bottom: 5rem;
-          right: 0;
-        `;
-      default:
-        return `
-          top: 3rem;
-          right: 0;
-        `;
-    }
-  }}
+  ${({ position }) => getPositionStyle(position || 'top-center')};
 `;
