+++ src/types/toast.ts
@@ -1,0 +1,5 @@
+export interface Toast {
+  id?: string;
+  content: string;
+  backgroundColor?: string;
+}
